requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
transformers
torch
python-dotenv

